Moved app routing to its own file
Better separation of concerns
